extracted segment getting classes from DashHandler.js
